Added nav drop-down and styled color fonts
For nav we added a drop down menu for the logged in features.  This
should help keep the menu bar looking free from clutter while providing
easy access to core features.  This is styled in a drop down / hover /
toggle combination so that it should work well on pc and mobile devices.
Also styled the colors and fonts to match the parchment / ink asthetic
we're going for.  This may need to be improved upon later, but it is
workable for now.
TODO: Die roller statistics and flesh out compaign view objects.